orginized into more files, not able to registers at_floor if a new button is pressed while inbetween floors
