+++ heislab/skeleton_project/source/driver/elevator.h
@@ -17,10 +17,9 @@
 #include <time.h>
 #include <signal.h>
 #include <unistd.h>
-#include <pthread.h>
-// #include "DestinationRequest.h"
-#include "request.h"
 #include "elevio.h"
+#include "utilities.h"
+#include "request.h"
 
 /**
  * @brief Defines an Elevator struct to store values.
@@ -67,39 +66,6 @@ void initialize_elevator(Elevator *elevator);
  * @param[in,out] elevator Pointer to the Elevator struct to be freed.
  */
 void free_elevator(Elevator *elevator);
-
-/**
- * @brief Listens for button presses and updates the elevator queue accordingly.
- *
- * This function runs in a separate thread, continuously checking for button presses.
- * When a button is pressed, it adds the corresponding request to the elevator's queue.
- *
- * @param arg Pointer to the Elevator structure.
- * @return Always returns NULL.
- */
-void *button_listener(void *arg);
-
-/**
- * @brief Monitors the elevator's current floor and handles stop conditions.
- *
- * This function runs in a separate thread, continuously checking if the elevator
- * has reached the correct floor and handling the necessary actions, such as stopping.
- *
- * @param arg Pointer to the Elevator structure.
- * @return Always returns NULL.
- */
-void *floor_listener(void *arg);
-
-/**
- * @brief Continuously checks for an emergency stop condition.
- *
- * This function runs in a separate thread and monitors if an emergency stop has been triggered.
- * If an emergency stop is detected, it handles stopping the elevator safely.
- *
- * @param arg Unused parameter (can be NULL).
- * @return Always returns NULL.
- */
-void *emergency_listener(void *arg);
 
 /**
  * @brief Initializes and runs the elevator control system.
@@ -233,10 +199,3 @@ void check_emergency_stop();
  * @param[in] elevator Pointer to the Elevator struct.
  */
 void print_elevator(Elevator *elevator);
-
-/**
- * @brief Runs a test for the sorting function.
- *
- * This function tests whether the elevator's queue sorting mechanism works correctly.
- */
-void test_elevator();
